SNAFU : Ne marche pas pour tous les problèmes, beaucoup trop lent Certains pb ne sont pas résolus correctement, si >20 : aucune solution trouvée

+++ src/Util.hs
@@ -1,7 +1,7 @@
 module Util where
 
 type ID = Int
-type ValType = Float
+type ValType = Double
 type Path = (ID, ID, Int)
 
 (#) :: (Read a) => String -> Int -> a
